zed create: Add -use flag
Add -use flag to zed create that if activated sets the created pool as
the current pool for future operations.
