+++ cmd/zed/create/command.go
@@ -5,6 +5,7 @@ import (
 	"flag"
 	"fmt"
 
+	"github.com/brimdata/zed/cli/poolflags"
 	"github.com/brimdata/zed/cmd/zed/root"
 	"github.com/brimdata/zed/lake/data"
 	"github.com/brimdata/zed/order"
@@ -40,6 +41,7 @@ type Command struct {
 	sortKey    string
 	thresh     units.Bytes
 	seekStride units.Bytes
+	use        bool
 }
 
 func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
@@ -48,9 +50,10 @@ func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
 		seekStride: units.Bytes(data.DefaultSeekStride),
 	}
 	c.thresh = data.DefaultThreshold
-	f.Var(&c.thresh, "S", "target size of pool data objects, as '10MB' or '4GiB', etc.")
-	f.StringVar(&c.sortKey, "orderby", "ts:desc", "comma-separated pool keys with optional :asc or :desc suffix to organize data in pool (cannot be changed)")
 	f.Var(&c.seekStride, "seekstride", "size of seek-index unit for ZNG data, as '32KB', '1MB', etc.")
+	f.Var(&c.thresh, "S", "target size of pool data objects, as '10MB' or '4GiB', etc.")
+	f.BoolVar(&c.use, "use", false, "set created pool as the current pool")
+	f.StringVar(&c.sortKey, "orderby", "ts:desc", "comma-separated pool keys with optional :asc or :desc suffix to organize data in pool (cannot be changed)")
 	return c, nil
 }
 
@@ -78,6 +81,12 @@ func (c *Command) Run(args []string) error {
 	}
 	if !c.LakeFlags.Quiet {
 		fmt.Printf("pool created: %s %s\n", poolName, id)
+	}
+	if c.use {
+		poolflags.WriteHead(poolName, "main")
+		if !c.LakeFlags.Quiet {
+			fmt.Printf("Switched to branch \"main\" on pool %q\n", poolName)
+		}
 	}
 	return nil
 }
