+++ cmd/zed/delete/command.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/brimdata/zed/cli/commitflags"
 	"github.com/brimdata/zed/cli/lakeflags"
+	"github.com/brimdata/zed/cli/poolflags"
 	"github.com/brimdata/zed/cmd/zed/root"
 	"github.com/brimdata/zed/lake/api"
 	"github.com/brimdata/zed/lakeparse"
@@ -43,12 +44,14 @@ can be "undone" by adding the commits back to the log using
 type Command struct {
 	*root.Command
 	commitFlags commitflags.Flags
+	poolFlags   poolflags.Flags
 	where       string
 }
 
 func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
 	c := &Command{Command: parent.(*root.Command)}
 	c.commitFlags.SetFlags(f)
+	c.poolFlags.SetFlags(f)
 	f.StringVar(&c.where, "where", "", "delete by pool key predicate")
 	return c, nil
 }
@@ -63,7 +66,7 @@ func (c *Command) Run(args []string) error {
 	if err != nil {
 		return err
 	}
-	head, err := c.LakeFlags.HEAD()
+	head, err := c.poolFlags.HEAD()
 	if err != nil {
 		return err
 	}
