+++ cmd/zed/merge/command.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/brimdata/zed/cli/commitflags"
 	"github.com/brimdata/zed/cli/lakeflags"
+	"github.com/brimdata/zed/cli/poolflags"
 	"github.com/brimdata/zed/cmd/zed/root"
 	"github.com/brimdata/zed/pkg/charm"
 )
@@ -23,12 +24,14 @@ var Cmd = &charm.Spec{
 type Command struct {
 	*root.Command
 	commitFlags commitflags.Flags
+	poolFlags   poolflags.Flags
 	force       bool
 }
 
 func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
 	c := &Command{Command: parent.(*root.Command)}
 	c.commitFlags.SetFlags(f)
+	c.poolFlags.SetFlags(f)
 	f.BoolVar(&c.force, "f", false, "force merge of main into a target")
 	return c, nil
 }
@@ -49,7 +52,7 @@ func (c *Command) Run(args []string) error {
 	if err != nil {
 		return err
 	}
-	head, err := c.LakeFlags.HEAD()
+	head, err := c.poolFlags.HEAD()
 	if err != nil {
 		return err
 	}
