+++ cmd/zed/use/command.go
@@ -5,7 +5,7 @@ import (
 	"flag"
 	"fmt"
 
-	"github.com/brimdata/zed/cli/lakeflags"
+	"github.com/brimdata/zed/cli/poolflags"
 	"github.com/brimdata/zed/cmd/zed/root"
 	"github.com/brimdata/zed/lakeparse"
 	"github.com/brimdata/zed/pkg/charm"
@@ -59,10 +59,13 @@ in the lake and is not copied to this local directory.
 
 type Command struct {
 	*root.Command
+	poolFlags poolflags.Flags
 }
 
 func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
-	return &Command{Command: parent.(*root.Command)}, nil
+	c := &Command{Command: parent.(*root.Command)}
+	c.poolFlags.SetFlags(f)
+	return c, nil
 }
 
 func (c *Command) Run(args []string) error {
@@ -75,7 +78,7 @@ func (c *Command) Run(args []string) error {
 		return errors.New("too many arguments")
 	}
 	if len(args) == 0 {
-		head, err := c.LakeFlags.HEAD()
+		head, err := c.poolFlags.HEAD()
 		if err != nil {
 			return errors.New("default pool and branch unset")
 		}
@@ -87,7 +90,7 @@ func (c *Command) Run(args []string) error {
 		return err
 	}
 	if commitish.Pool == "" {
-		head, err := c.LakeFlags.HEAD()
+		head, err := c.poolFlags.HEAD()
 		if err != nil {
 			return errors.New("default pool unset")
 		}
@@ -110,7 +113,7 @@ func (c *Command) Run(args []string) error {
 	if _, err = lake.CommitObject(ctx, poolID, commitish.Branch); err != nil {
 		return err
 	}
-	if err := lakeflags.WriteHead(commitish.Pool, commitish.Branch); err != nil {
+	if err := poolflags.WriteHead(commitish.Pool, commitish.Branch); err != nil {
 		return err
 	}
 	if !c.LakeFlags.Quiet {
