+++ lake/ztests/compact.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   seq 10 | zq '{ts:this-1,s:"val${this-1}"}' - | zed load -q -
   seq 10 | zq '{ts:this-1,s:"val${this-1}"}' - | zed load -q -
   ids=$(zed query -f text 'from test@main:objects | yield "0x${hex(id)}"')
