+++ lake/ztests/delete-where-invalid.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q test
-  zed use -q test 
+  zed create -use -q test
   ! zed delete -where 'me == "hi" | cut me'
   ! zed delete -where 'cut that'
