+++ lake/ztests/delete-where-missing.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   zed load -q in.zson
   ! zed delete -q -where 'uid==C3UeSqaSOFRReHD68'
   zed query -z 'count:=count()'
