+++ lake/ztests/delete-where.yaml
@@ -3,8 +3,7 @@ script: |
   zed init -q
   for order in desc asc; do
     echo === $order ===
-    zed create -q -S 1KB -orderby ts:$order test
-    zed use -q test
+    zed create -use -q -S 1KB -orderby ts:$order test
     seq 1000 | zq '{ts:this-1,s:"val${this-1}"}' - | zed load -q -
     zed delete -q -where 'ts > 400 and ts <= 500'
     zed query -z 'count:=count()'
