+++ lake/ztests/delete.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q -orderby x:desc test
-  zed use -q test
+  zed create -use -q -orderby x:desc test
   zed load -q 1.zson
   id=$(zed query -f text "from test@main:objects | cut id:=ksuid(id) | tail 1")
   zed load -q 2.zson
