+++ lake/ztests/meta-optimized-filter.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   seq 20 | zq 'yield {ts: this}' - | zed load -q -
   seq 21 40 | zq 'yield {ts: this}' - | zed load -q -
   zed query -z 'from test:objects | max > 20 | cut min, max'
