+++ lake/ztests/overlap.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q logs
-  zed use -q logs
+  zed create -use -q logs
   zed load -q babble-split1.zson
   zed load -q babble-split2.zson
   zed query -Z "from logs@main:objects | sort -r size | drop id"
