+++ lake/ztests/revert.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   a=$(zed load a.zson | head -1 | awk '{print $1}')
   b=$(zed load b.zson | head -1 | awk '{print $1}')
   zed query -z "sort this"
