+++ lake/ztests/time-travel.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q POOL
-  zed use -q POOL
+  zed create -use -q POOL
   a=$(zed load a.zson | head -1 | awk '{print $1}')
   b=$(zed load b.zson | head -1 | awk '{print $1}')
   zed query -z "from POOL | sort this"
