+++ lake/ztests/vacuum.yaml
@@ -1,8 +1,7 @@
 script: |
   export ZED_LAKE=test
   zed init -q
-  zed create -q -orderby x:asc test
-  zed use -q test
+  zed create -use -q -orderby x:asc test
   echo {x:1} | zed load -q -
   r=$(echo {x:2} | zed load - | head -1 | awk '{print $1}')
   r=$(zed revert $r | awk '{print $5}')
