+++ service/ztests/default-format.yaml
@@ -1,8 +1,7 @@
 script: |
   export LAKE_EXTRA_FLAGS='-defaultfmt=ndjson'
   source service.sh
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   echo '{x: 1(uint64)}' | zed load -q -
   curl -d '{"query": "from test"}' $ZED_LAKE/query
