+++ service/ztests/delete-where.yaml
@@ -2,8 +2,7 @@ script: |
   source service.sh
   for order in desc asc; do
     echo === $order === | tee >(cat >&2)
-    zed create -q -S 1KB -orderby ts:$order test
-    zed use -q test
+    zed create -use -q -S 1KB -orderby ts:$order test
     seq 1000 | zq '{ts:this-1,s:"val${this-1}"}' - | zed load -q -
     zed delete -q -where 'ts > 400 and ts <= 500'
     zed query -z 'count()'
