+++ service/ztests/log.yaml
@@ -1,7 +1,6 @@
 script: |
   source service.sh
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   zed load -user testuser -q 1.zson
   zed load -user testuser -q 2.zson
   zed log
