+++ service/ztests/query-sql-agg.yaml
@@ -2,8 +2,7 @@
 
 script: |
   source service.sh
-  zed create -q test
-  zed use -q test
+  zed create -use -q test
   zed load -q data.zson
   zed query -z 'SELECT count(), id GROUP BY id | sort id'
